requirements.txt
langchain>=0.1.0
langchain_anthropic>=0.1.0
langchain_community>=0.1.0
chromadb>=0.4.22
streamlit>=1.28.0
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pypdf>=5.4.0
python-docx>=0.8.11
docx2txt>=0.8
easyocr>=1.7.0
Pillow>=9.0.0
hf_xet>=1.0.5